Move Positioned and Line into common/ subfolder
