+++ ui/src/components/Editor/EditorSvg/SvgDragLayer.tsx
@@ -1,7 +1,7 @@
 import { getItemWithKind, useDragLayer } from "util/dnd";
 import { useDOMToSVGPosition } from "util/svg";
 import DragPreview from "./SvgDragPreview";
-import Positioned from "./Positioned";
+import Positioned from "./common/Positioned";
 
 const SvgDragLayer: React.FC = () => {
   const { itemWithKind, currentOffset } = useDragLayer((monitor) => ({
